Fix create-commit.sh script to use multi-line format by default
Prompt: there is a problem with the create-commit script. the last commit message looks to be what should have been the commit description. fix the issue

Description:
Modified create-commit.sh to:
- Use multi-line format by default instead of single-line
- Change the --multi-line flag to --single-line since multi-line is now the default
- Update documentation and help text to reflect these changes

Discussion:
The main issue was that the script was concatenating all parts of the commit message into a single line, making it difficult to read. The yastai.mdc guidelines in CONTRIBUTING.md show that proper multi-line commit messages are the intended format.