Add AI PR helper script
Prompt: this pr was perfect, can you save everything you did to achieve it ? I want you to be able to easily replicate this process. Maybe update the create-pr script or update the ai instructions.

Description:
Created a new AI-friendly script for pull request creation

Discussion:
No major challenges were faced. The main task was identifying the source of the multiline command issue and creating a specialized script to work around it.